requirements.txt
gradio
transformers
torch
Pillow
fastapi
